requirement.txt
tensorflow==1.15.2
keras==2.2.4
Pillow==6.2.1
aiofiles==0.4.0
aiohttp==3.6.2
starlette==0.13.0
uvicorn==0.10.8
python-multipart==0.0.5
opencv-python==4.1.2.30